Adding mockup for mutations with fragments
In order to mock mutations with fragments and use MockedProvider, it is required to add the "__typename" field in our mocks.
Also, if the response payload does not contain a field named "id" or "_id", we need to implement a custom InMemoryCache to use one of the fields as the id from our response object.